+++ app/__init__.py
@@ -1,12 +1,15 @@
 from flask import Flask
 import os
+from werkzeug.middleware.proxy_fix import ProxyFix
 
 
 def create_app():
-    app = Flask(__name__)
+    app = Flask(__name__)    
 
     CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
     CONFIG_PATH = os.path.join(CURRENT_DIR, '..', 'config.py')
+
+    app.wsgi_app = ProxyFix(app.wsgi_app, x_for=1, x_proto=1, x_host=1, x_port=1, x_prefix=1)
     
     config_namespace = {}
     with open(CONFIG_PATH, 'r') as file: